requirements.txt
amqp==2.4.1
billiard==3.5.0.5
cachetools==3.1.0
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
cloudinary==1.15.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
coveralls==1.5.1
dj-database-url==0.5.0
Django==2.1.5
django-celery-results==1.0.4
django-cloudinary-storage==0.2.3
django-cors-middleware==1.3.1
django-extensions==2.1.4
django-filter==2.1.0
djangorestframework==3.9.1
docopt==0.6.2
drf-yasg==1.13.0
facebook-sdk==3.1.0
future==0.17.1
google-auth==1.6.2
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
kombu==4.3.0
mailer==0.8.1
MarkupSafe==1.1.0
mock==2.0.0
oauthlib==3.0.1
pbr==5.1.2
Pillow==5.4.1
psycopg2==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
PyJWT==1.7.1
python-dotenv==0.10.1
python-twitter==3.5
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
ruamel.yaml==0.15.87
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
whitenoise==4.1.2
beautifulsoup4==4.7.1
cssselect==1.0.3
lxml==4.3.1
markdown2==2.3.7
pyquery==1.4.0
readtime==1.0.6
soupsieve==1.7.3
